bug fix images were binary because I was pulling uint16 (line 62) and treating it as float32 (line 136)

+++ main.py
@@ -59,7 +59,7 @@ evalscript = """
 function setup() {
   return {
     input: ["B02", "B04", "B08", "B11"],
-    output: { bands: 4, sampleType: "UINT16" }
+    output: { bands: 4, sampleType: "FLOAT32" }
   };
 }
 function evaluatePixel(sample) {
